Pass global flags down to build command (2nd try) (#152)
* Fix global flag parsing
