+++ bin/makisu/cmd/build.go
@@ -101,7 +101,6 @@ var (
 		DisableFlagsInUseLine: true,
 		Short: "Build docker image, optionally push to registries and/or load into docker daemon",
 
-		TraverseChildren: true,
 		Args: func(cmd *cobra.Command, args []string) error {
 			if len(args) != 1 {
 				return errors.New("Requires build context as argument")
@@ -109,6 +108,18 @@ var (
 			return nil
 		},
 		Run: func(cmd *cobra.Command, args []string) {
+			if cleanup, err := processGlobalFlags(); err != nil {
+				fmt.Println(err)
+				os.Exit(1)
+			} else {
+				defer cleanup()
+			}
+
+			if err := processFlags(); err != nil {
+				log.Errorf("failed to process flags: %s", err)
+				os.Exit(1)
+			}
+
 			if err := Build(args[0]); err != nil {
 				log.Error(err)
 				os.Exit(1)
@@ -194,10 +205,6 @@ func newBuildPlan(
 // If --load is specified, will load the image into the local docker daemon.
 func Build(contextDir string) error {
 	log.Infof("Starting Makisu build (version=%s)", utils.BuildHash)
-
-	if err := processFlags(); err != nil {
-		return fmt.Errorf("failed to process flags: %s", err)
-	}
 
 	// Create BuildContext.
 	contextDirAbs, err := filepath.Abs(contextDir)
