+++ bin/makisu/cmd/version.go
@@ -30,7 +30,6 @@ var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Print version number",
 
-	TraverseChildren: true,
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println(utils.BuildHash)
 	},
